# optimize layer norm backward with hidden size in the range of 7K to 10K by increasing maximum allowed batch size by 1 and changing threads_per_block searching method