**Issue:**
The performance dropped from 1400 GB/s to 1100 GB/s when the hidden size increased from 7K to 7K+128. This is because at 7K, vect = 8, threads_per_block = 128, and batch = 7. The maximum allowed batch size without causing register spill is 7. So at hidden size 7K+128, threads_per_block is increased to 160 and batch is decreased to 6. Blocks per SM is reduced from 2 to 1.
![image](https://github.com/NVIDIA/Fuser/assets/116412316/cb4f3867-8b84-4a5a-83bf-50a6f394e780)


**Fix:**

1. Allow a small register spill by increasing the maximum allowed batch size from 7 to 8. With this fix, the bandwidth can maintain around 1350 GB/s for cases with hidden size in the range of (7K, 8K]. But the bandwidth still drops to 1100 GB/s after 8K. Tried to further increase maximum batch size to 9, the bandwidth still stays around 1100 GB/s.
![image](https://github.com/NVIDIA/Fuser/assets/116412316/89990712-67c1-4317-b984-b0dc83339659)

2. Double threads_per_block instead of increasing by 32. In the original code, when batch size exceeds the maximum allowed value, threads_per_block will be increased by a step_size and check if the new batch size is smaller than maximum allowed value. With this searching method, the next threads_per_block is 160 and blocks per SM is reduced from 2 to 1. This means the code is only using 160 x 255 registers instead of all the 64K registers on each SM. To reduce this register waste, we can directly double threads_per_block from 128 to 256 and still have 255 registers per thread and one block per SM.
![image](https://github.com/NVIDIA/Fuser/assets/116412316/6c4339da-3db1-406c-a8e4-930de5aec907)

**Results:**
Bandwidth increased about 10% for cases with hidden size in the range of (7K,8K]
Bandwidth increased about 5% for cases with hidden size in the range of (8K,10K]
![image](https://github.com/NVIDIA/Fuser/assets/116412316/11cb6436-5cc5-4ce5-8ac6-5dd962e7ee0a)