@naoyam 2023-07-31
Did the last graph use the same problem configuration? It shows much better bandwidths across the hidden sizes