@liqiangxl 2023-07-31
> Did the last graph use the same problem configuration? It shows much better bandwidths across the hidden sizes

Only the bandwidths in the last graph are real bandwidths. All others are scaled to [0,40] so they can be plot on the heuristic para graph.