@naoyam 2023-08-03
> Double threads_per_block instead of increasing by 32. In the original code, when batch size exceeds the maximum allowed value, threads_per_block will be increased by a step_size and check if the new batch size is smaller than maximum allowed value. With this searching method, the next threads_per_block is 160 and blocks per SM is reduced from 2 to 1. This means the code is only using 160 x 255 registers instead of all the 64K registers on each SM. To reduce this register waste, we can directly double threads_per_block from 128 to 256 and still have 255 registers per thread and one block per SM.

This seems to depend on the number of blocks per SM. If it's just 1, the minimum number of threads should be 256 if the register usage is the most important. If it's 2, 128 threads should be enough. shouldn't we increment the thread count until the minimum is met instead of just always doubling it?