@naoyam 2023-08-03
Did you also measure the performance with just the second optimization? I wonder how much impact the fist optimization actually has.