@liqiangxl 2023-08-03
> Did you also measure the performance with just the second optimization? I wonder how much impact the fist optimization actually has.

The first optimization only influence cases with hidden size in the range of (7K,8K] and the bandwidth increased about 10% of hardward theoretical bandwidth.