@liqiangxl 2023-08-03
> > Double threads_per_block instead of increasing by 32. In the original code, when batch size exceeds the maximum allowed value, threads_per_block will be increased by a step_size and check if the new batch size is smaller than maximum allowed value. With this searching method, the next threads_per_block is 160 and blocks per SM is reduced from 2 to 1. This means the code is only using 160 x 255 registers instead of all the 64K registers on each SM. To reduce this register waste, we can directly double threads_per_block from 128 to 256 and still have 255 registers per thread and one block per SM.
> 
> This seems to depend on the number of blocks per SM. If it's just 1, the minimum number of threads should be 256 if the register usage is the most important. If it's 2, 128 threads should be enough. shouldn't we increment the thread count until the minimum is met instead of just always doubling it?

As you said, with 2 blocks per SM we need 128 threads. Then the next value is 1 blocks per SM, which needs 256 threads. There is no intermediate values we need to increment slowly.
The hardware has 64K registers, and this can be broken down in several ways: 64K equals 1 x 256 threads x 256 registers per thread, or 2 x 128 x 256, or 4 x 64 x 256, or 8 x 32 x 256. Mathematically, since 256 is equivalent to 2^8, consistently doubling the number of threads ensures that all of the registers are fully utilized. Therefore, such a strategy is safe for making complete use of the available registers.
Due to the change of persistent batch size, each threads will not always use 255 registers. But for this case, we only need to increase threads per block when persistent batch size is larger than 8. So the smallest batch is ceilDiv(9,2) = 5, which will use 216 registers.