@naoyam 2023-08-03
> As you said, with 2 blocks per SM we need 128 threads. Then the next value is 1 blocks per SM, which needs 256 threads. There is no intermediate values we need to increment slowly.

I was thinking about the case like:

> With this searching method, the next threads_per_block is 160 and blocks per SM is reduced from 2 to 1.

I assume we start with 128 threads and then try 160 threads. If the occupancy drops to 1, then we should jump to 256 threads. But what if not? If the occupancy stays 2, then the SM register space would still be fully used, right?