@liqiangxl 2023-08-04
> > > Did you also measure the performance with just the second optimization? I wonder how much impact the fist optimization actually has.
> > 
> > 
> > We will lose 10% bandwidth from 7K to 8K if use only the second optimization compared to use both 1st and 2nd optimization. ![image](https://user-images.githubusercontent.com/116412316/258149992-51373216-c0c9-4d25-89a9-0e6aaabda76f.png)
> 
> Thanks. My concern is this seems to be highly sensitive to the register usage, which could vary between compiler versions as well as different target architectures. What does the performance look like on H100 and V100?


Here is V100. From 7K to 8K, the bandwidth dropped about 3%. V100 has a smaller L1 caches, so the negative effect of register spill is larger the benefit it brings.
![image](https://github.com/NVIDIA/Fuser/assets/116412316/edd0609e-5a7d-4487-abf6-d786bfab59db)
 Here is H100 (copied from #666, the performanc change in the block is due to this PR.):
![image](https://github.com/NVIDIA/Fuser/assets/116412316/92e0407f-214d-4360-bbd6-93667594a330)