@liqiangxl 2023-08-04
> I assume we start with 128 threads and then try 160 threads. If the occupancy drops to 1, then we should jump to 256 threads. But what if not? If the occupancy stays 2, then the SM register space would still be fully used, right?

The problem here is 2 x 128 threads x 256 registers per thread = 64K, any increase of threads count will drop blocks per sm from 2 to 1. Unless we limit the register count to a much smaller value, e.g.  for 160 threads, to achieve 2 blocks per sm each thread can only use up to 168 registers, which is too small (a batch size of 5 needs 216 registers). This also can't fully use the registers. To fully use registers, the avilable block sizes are 64, 128, 256, ...