@naoyam 2023-08-04
OK, makes sense with this:

> a batch size of 5 needs 216 registers). 

Can you please add more detailed comments?