@liqiangxl 2023-08-05
> OK, makes sense with this:
> 
> > a batch size of 5 needs 216 registers).
> 
> Can you please add more detailed comments?

updated.