@naoyam 2023-08-07
I wonder if we should use shmem in these sizes too instead of speculating register spill performance.