@liqiangxl 2023-08-07
> I wonder if we should use shmem in these sizes too instead of speculating register spill performance.

I conducted a preliminary test by enforcing the use of smem at hidden size of 8K. As a result, the kernel runtime experienced a minor increase, moving from 1.158 ms to 1.185 ms. For a comprehensive test, we'd need significant modifications and further optimization of the heuristics, since the persistent grid size could also be constrained by the shared memory. In the method currently in place in #666, shared memory is only tapped into when registers are insufficient for persistent buffers, with occupancy consistently restricted by registers. Perhaps it would be prudent to consider this advanced optimization or testing after the completion of #666. It's akin to first establishing the capability and then determining how to derive the greatest value from it.